Remove duplicates and inherits from base class  - NinchatConversationQuestionnaire  - NinchatFormQuestionnaire
